requirements.txt
awscli==1.37.21
backports.tarfile==1.2.0
blinker==1.9.0
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker==7.1.0
docutils==0.16
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.5
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.75
ultralytics-thop==2.0.14
urllib3==2.3.0
voluptuous==0.15.2
Werkzeug==3.1.3
yolo==0.3.1
